+++ packages/curd-table/src/components/TableEditItem.vue
@@ -147,7 +147,7 @@
     <!-- upload -->
     <np-custom-upload
       v-if="type === 'upload'"
-      v-model:value="formValue"
+      v-model:value="customFileList"
       :label="optionConfig?.label"
       :accept="optionConfig?.accept"
       :info="info"
@@ -252,6 +252,20 @@ const dateRangeValue = computed<[string, string] | null>({
       props.form[props.option?.config?.startField || 'start'] = v[0];
       props.form[props.option?.config?.endField || 'end'] = v[1];
     }
+  },
+});
+
+// Upload
+const customFileList = computed({
+  get() {
+    return props.option?.config?.parse
+      ? props.option?.config?.parse.get(formValue.value)
+      : formValue.value;
+  },
+  set(v) {
+    formValue.value = props.option?.config?.parse
+      ? props.option?.config?.parse.set(v)
+      : v;
   },
 });
 </script>